+++ module-ticket/src/main/java/com/example/moduleticket/domain/ticket/service/TicketService.java
@@ -7,17 +7,16 @@ import static com.example.modulecommon.exception.ErrorCode.USER_ACCESS_DENIED;
 import com.example.modulecommon.entity.AuthUser;
 import com.example.modulecommon.exception.ServerException;
 import com.example.moduleticket.domain.reservation.entity.Reservation;
-import com.example.moduleticket.domain.ticket.dto.TicketContext;
 import com.example.moduleticket.domain.ticket.dto.RefundDto;
+import com.example.moduleticket.domain.ticket.dto.TicketContext;
 import com.example.moduleticket.domain.ticket.dto.response.TicketResponse;
 import com.example.moduleticket.domain.ticket.entity.Ticket;
 import com.example.moduleticket.domain.ticket.repository.TicketRepository;
 import com.example.moduleticket.feign.GameClient;
 import com.example.moduleticket.feign.PaymentClient;
+import com.example.moduleticket.feign.SeatClient;
 import com.example.moduleticket.feign.dto.GameDto;
 import com.example.moduleticket.feign.dto.SeatDto;
-import com.example.moduleticket.feign.dto.request.PaymentRequest;
-import com.example.moduleticket.util.IdempotencyKeyUtil;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
@@ -38,6 +37,7 @@ public class TicketService {
 	private final TicketPaymentService ticketPaymentService;
 	private final TicketCreateService ticketCreateService;
 	private final GameClient gameClient;
+	private final SeatClient seatClient;
 	private final PaymentClient paymentClient;
 
 	@Transactional(readOnly = true)
@@ -73,11 +73,18 @@ public class TicketService {
 
 
 	@Transactional
-	public TicketResponse issueTicketFromReservation(AuthUser auth, GameDto gameDto, List<SeatDto> seats,
+	public TicketResponse issueTicketFromReservation(AuthUser auth, Long gameId, List<Long> seatIds,
 		Reservation reservation) {
 
-		TicketContext ticketContext = ticketCreateService.createTicket(auth, gameDto, seats, reservation);
-		ticketPaymentService.paymentTicket(ticketContext);
+		GameDto gameDto = gameClient.getGame(gameId);
+		List<SeatDto> seatDtos = seatClient.getSeatsByGame(gameId, seatIds);
+
+		TicketContext ticketContext = ticketCreateService.createTicket(auth, gameDto, seatDtos, reservation);
+		ticketPaymentService.create(
+			ticketContext.getTicket(),
+			ticketContext.getMemberId(),
+			ticketContext.getTotalPoint()
+		);
 
 		return ticketContext.toResponse();
 	}
